# infrastructure improvements make off_t pointer-sized (required for >4gb files at work) add shared_ptr deleter for free(), add ASSUME_ALIGNED, warning fixes
git-svn-id: http://svn.wildfiregames.com/public/ps/trunk@6826 3db68df2-c116-0410-a063-a993310a9797

+++ source/lib/sysdep/os/win/wnuma.cpp
@@ -326,7 +326,7 @@ static bool VerifyPages(void* mem, size_t size, size_t pageSize, size_t node)
 	PSAPI_WORKING_SET_EX_INFORMATION* wsi = new PSAPI_WORKING_SET_EX_INFORMATION[numPages];
 	for(size_t i = 0; i < numPages; i++)
 		wsi[i].VirtualAddress = (u8*)mem + i*pageSize;
-	pQueryWorkingSetEx(GetCurrentProcess(), wsi, sizeof(PSAPI_WORKING_SET_EX_INFORMATION)*numPages);
+	pQueryWorkingSetEx(GetCurrentProcess(), wsi, DWORD(sizeof(PSAPI_WORKING_SET_EX_INFORMATION)*numPages));
 
 	// ensure each is valid and allocated on the correct node
 	for(size_t i = 0; i < numPages; i++)
